Add Files via Upload
This is the code for the Reviews v Best Sellers.  I can discuss what I did more tomorrow night, but the long and short is that I dropped anything with zero reviews and then analyzed the remaining data.  Prior to that, I did capture the top 25 categories that had best sellers and no reviews.  I did not delve further into that yet as I haven't gotten an answer on Slack.  I can see if i can gain any qualitative insight from those categories.  There are about 807000 products with no reviews.  There are 5426 out of that that are best sellers.  We can definitely use this data.  Once dropped, their was a high (.71) correlation between reviews and best sellers.  I have three visualizations for the presentation that tell this story.  they are uploaded with this commit.